backend/requirements.txt
fastapi>=0.104.0
uvicorn>=0.24.0
sqlalchemy>=2.0.23
aiosqlite>=0.19.0
greenlet>=3.0.0
pydantic>=2.5.0
pydantic-settings>=2.1.0
python-dotenv>=1.0.0
httpx>=0.25.0
aiohttp>=3.9.0
arxiv>=2.1.0
PyMuPDF>=1.24.0
anthropic>=0.39.0
apscheduler>=3.10.0
python-multipart>=0.0.6
pyyaml>=6.0
jinja2>=3.1.0

# 验证脚本依赖
playwright>=1.40.0

# 测试依赖
pytest>=8.0.0
pytest-asyncio>=0.23.0

# MCP Server 依赖
sse-starlette>=2.0.0